% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalise_plot.R
\name{build_release_charts_doc}
\alias{build_release_charts_doc}
\title{Function to create an Excel workbook to save charts to}
\usage{
build_release_charts_doc(workbook_name)
}
\arguments{
\item{workbook_name}{The name of the excel workbook to create}
}
\value{
initialised Excel workbook
}
\description{
Function to create an Excel workbook to save charts to
}
\examples{
workbook <- build_release_charts_doc("charts")
}
