% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalise_plot.R
\name{chart_folder_to_workbook}
\alias{chart_folder_to_workbook}
\title{Function to add chart to a release file with all charts in}
\usage{
chart_folder_to_workbook(
  project_area,
  workbook_name,
  chart_names_list,
  titles_list,
  sources_list,
  alt_texts_list
)
}
\arguments{
\item{project_area}{The filepath to save the work in a charts folder (for
charts saved in current directory use ".")}

\item{workbook_name}{Name of Excel workbook to save charts in}

\item{chart_names_list}{List of the chart names (to use as Tab name)}

\item{titles_list}{List of titles use as chart titles in order}

\item{sources_list}{List of chart sources in order}

\item{alt_texts_list}{List of alt text in order for the plots}
}
\description{
Function to add chart to a release file with all charts in
}
\examples{

colors <- color_picker(3)
example_chart <- ggplot2::ggplot(data = iris) +
  ggplot2::aes(x = Petal.Length, y = Petal.Width) +
  ggplot2::geom_point(ggplot2::aes(color = Species, shape = Species)) +
  kas_style()
dir.create("./charts")
finalise_plot("chart_six", example_chart, "./charts/")

title_one <- "Petal length and width by species"
chart_path_one <- "./charts/chart_six.svg"
source_one <- "Iris"
alt_one <- "Petal length and width has a positive association"
chart_folder_to_workbook(
  "./", "charts.xlsx", c("chart_six"),
  c(title_one), c(source_one), c(alt_one)
)
}
