% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_picker.R
\name{color_picker}
\alias{color_picker}
\title{Function returns a list of colors to use in a chart bound to a category. With
error handling to warn users if too many colours are
being used with a certain
palette.}
\usage{
color_picker(categories, primary_palette = FALSE)
}
\arguments{
\item{categories}{The list of categories in to assign colors to for the chart}

\item{primary_palette}{Boolean indicator of whether to use the primary colour
palette or not (T/F)}
}
\value{
list of colours matched to a category
}
\description{
Function returns a list of colors to use in a chart bound to a category. With
error handling to warn users if too many colours are
being used with a certain
palette.
}
\examples{
color_picker(3)
color_picker(c("series1", "series2", "series3"), primary_palette = TRUE)
}
