% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalise_plot.R
\name{finalise_plot}
\alias{finalise_plot}
\title{Finalising plots for release, the finalise plot function alone will save a
plot in a safe size, with the extension being SVG for accessibility/usability}
\usage{
finalise_plot(name, plot, path, width = 17.6, height = 16.5)
}
\arguments{
\item{name}{The plot name to save as (no file extension needed)}

\item{plot}{The plot to save}

\item{path}{The file path to save the plot on}

\item{width}{The width of the chart}

\item{height}{The height of the chart}
}
\description{
Finalising plots for release, the finalise plot function alone will save a
plot in a safe size, with the extension being SVG for accessibility/usability
}
\examples{
example_chart <- ggplot2::ggplot(data = iris) +
  ggplot2::aes(x = Petal.Length, y = Petal.Width) +
  ggplot2::geom_point(ggplot2::aes(color = Species, shape = Species))

dir.create("./charts")
finalise_plot("chart_six", example_chart, "./charts/")
}
