% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kas_style.R
\name{kas_style}
\alias{kas_style}
\title{Add KAS theme to ggplot chart}
\usage{
kas_style()
}
\description{
This function allows you to add the KAS theme to your ggplot graphics.
}
\examples{
example_chart <- ggplot2::ggplot(data = iris) +
  ggplot2::aes(x = Petal.Length, y = Petal.Width) +
  ggplot2::geom_point(ggplot2::aes(color = Species, shape = Species)) +
  kas_style()
}
\keyword{kas_style}
